Fetch User Data + Minor Changes
- Changes to the connection objects to gRPC services to support any stub
- Minor change to the URL prefix's for the gateway service
- Moved proto message conversion funcs to seperate file
- Verification config file created for new RPC calls planned
- Middleware to check whether the user is logged in and has valid and accurate session data

Stubs must be included when using ServerCommunication.grpc_request as it is the best and cleanest solution.
Alternative was to store and loop through all stubs which wastes time and resources. Would only be useful if there was a way to add stubs and extract all RPC functions from the class to accurately use the correct call.