Customer Account Settings (#2)
* Update README.md

* Created User Action Service

- Created new .proto file
- Basic file for user action service created
- Added to server

* Separated Functions To New File

* Fetch User Data + Minor Changes

- Changes to the connection objects to gRPC services to support any stub
- Minor change to the URL prefix's for the gateway service
- Moved proto message conversion funcs to seperate file
- Verification config file created for new RPC calls planned
- Middleware to check whether the user is logged in and has valid and accurate session data

Stubs must be included when using ServerCommunication.grpc_request as it is the best and cleanest solution.
Alternative was to store and loop through all stubs which wastes time and resources. Would only be useful if there was a way to add stubs and extract all RPC functions from the class to accurately use the correct call.

* All Gateway Routes Created

There is a current error with my implementation. I am using a user cookie that has the same data as the Redis session. Will change to not use a user cookie (should not be created anyways).

* Changed Gateway To Only Use Session Cookie Data

User data is fetched by the Redis session which only requires the client to provide the user uuid

* Change Account Email Functional

When account email is changed, the user is sent another OTP verification code to re-verify their account. They are unable to use their account until it has been verified.

* Change Account Password Complete

- Minor fix on login attempts crash

* Change Account Details

- Added in skip for data validation if the data is blank but not None

* Account Deletion Functional

* Documentation Update

Completed the docstrings of various functions that were either incomplete or empty.

* Added Postman Collection

* Update README.md